Style Update 분리
